Schreibe ServerGame und ClientGame

+++ gdx-bunny-brawl/src/main/java/de/hochschuletrier/gdw/ws1314/entity/ServerEntityManager.java
@@ -51,7 +51,7 @@ public class ServerEntityManager {
     }
 
 
-    public void update(int delta) {
+    public void update(float delta) {
         internalRemove();
         internalInsert();
